+++ packages/access-definitions/src/Wrapping/Utilities/SchemaPathProcessor.php
@@ -65,43 +65,6 @@ class SchemaPathProcessor
     /**
      * Check the paths of the given sort methods for availability and aliases using the given type.
      *
-     * If no sort methods were given then apply the {@link TypeInterface::getDefaultSortMethods() default sort methods}
-     * of the given type.
-     *
-     * @template TCondition of \EDT\Querying\Contracts\PathsBasedInterface
-     * @template TSorting of \EDT\Querying\Contracts\PathsBasedInterface
-     *
-     * @param TypeInterface<TCondition, TSorting, object> $type
-     * @param TSorting                           ...$sortMethods
-     *
-     * @return list<TSorting>
-     *
-     * @throws AccessException
-     * @throws PathException Thrown if {@link TypeInterface::getAliases()} returned an invalid path.
-     *
-     * @deprecated use {@link SchemaPathProcessor::processDefaultSortMethods()} and {@link SchemaPathProcessor::mapSorting()} instead
-     */
-    public function mapSortMethods(TypeInterface $type, PathsBasedInterface ...$sortMethods): array
-    {
-        if ([] === $sortMethods) {
-            return $this->processDefaultSortMethods($type);
-        }
-
-        if ($type instanceof SortableTypeInterface) {
-            $typeAccessor = new ExternSortableTypeAccessor($this->typeProvider);
-            $processor = $this->propertyPathProcessorFactory->createPropertyPathProcessor($typeAccessor);
-            foreach ($sortMethods as $sortMethod) {
-                $processor->processPropertyPaths($sortMethod, $type);
-            }
-            return array_values($sortMethods);
-        }
-
-        throw AccessException::typeNotSortable($type);
-    }
-
-    /**
-     * Check the paths of the given sort methods for availability and aliases using the given type.
-     *
      * @template TCondition of \EDT\Querying\Contracts\PathsBasedInterface
      * @template TSorting of \EDT\Querying\Contracts\PathsBasedInterface
      *